[资源中心/非结构化] 完善数据资产区与资源中心闭环，并落地非结构化文件/元数据/清洗
补齐业务库种子与十大模块盘点，增强分区/存储/目录/监控；同步非结构化治理与 Flyway V138～V145。

+++ platform-backend/src/main/java/com/chengde/smartcity/masterdata/service/MetadataSubsystemService.java
@@ -1724,8 +1724,8 @@ public class MetadataSubsystemService {
             }
             items.add(row);
         }
-        unreadNotice = (int) changeNoticeMapper.selectCount(new LambdaQueryWrapper<GovMetaChangeNotice>()
-                .eq(GovMetaChangeNotice::getStatus, "UNREAD"));
+        unreadNotice = Math.toIntExact(changeNoticeMapper.selectCount(new LambdaQueryWrapper<GovMetaChangeNotice>()
+                .eq(GovMetaChangeNotice::getStatus, "UNREAD")));
         matchedHint = (int) entries.stream().filter(e -> e.getOmRef() != null && !e.getOmRef().isBlank()).count();
         Map<String, Object> kpi = new LinkedHashMap<>();
         kpi.put("total", entries.size());
